anjay slebew mabar yhahahayuk

+++ src/katalog/katalog.service.js
@@ -3,15 +3,15 @@ const {
   insertKatalog,
   updateKatalog,
   deleteKatalog,
-  findKatalogs,
-  findKatalogById,
+  findKatalogByProductCode,
 } = require("./katalog.repository.js");
 const codeGenerator = require("../helper/codegenerator.js");
 const { removeFile } = require("../helper/fileRemover.js");
+const { datePicker } = require("../helper/datePicker.js");
 
-async function getKatalogs() {
+async function getKatalogs(nama_produk) {
   try {
-    const users = await showKatalogs();
+    const users = await showKatalogs(nama_produk);
     return users;
   } catch (error) {
     throw error;
@@ -31,9 +31,11 @@ async function createKatalog(formData) {
 
 async function editKatalog(formData, id) {
   try {
-    const results = await getKatalogById(id);
-    let filePath = "./public/product_images/" + results.foto_produk;
-    removeFile(filePath);
+    const results = await getKatalogByProductCode(id);
+    if (formData.foto_produk){
+      let filePath = "./public/product_images/" + results.foto_produk;
+      removeFile(filePath);
+    }
     if (results) {
       const users = await updateKatalog(formData, id);
       return users;
@@ -43,32 +45,26 @@ async function editKatalog(formData, id) {
   }
 }
 
-async function removeKatalog(id) {
+async function removeKatalog(kode_produk) {
   try {
-    const results = await getKatalogById(id);
-    let filePath = "./public/product_images/" + results.foto_produk;
-    removeFile(filePath);
-    if (results) {
-      const users = await deleteKatalog(id);
-      return users;
+    const results = await getKatalogByProductCode(kode_produk);
+    if (results){
+      const deleted_date = datePicker();
+      const users = await deleteKatalog(deleted_date, kode_produk);
+      return true;
     }
+    return false;
+    // let filePath = "./public/product_images/" + results.foto_produk;
+    // removeFile(filePath);
+      
   } catch (error) {
     throw error;
   }
 }
 
-async function searchKatalog(keywords) {
+async function getKatalogByProductCode(kode_produk) {
   try {
-    const users = await findKatalogs(keywords);
-    return users;
-  } catch (error) {
-    throw error;
-  }
-}
-
-async function getKatalogById(id) {
-  try {
-    const results = await findKatalogById(id);
+    const results = await findKatalogByProductCode(kode_produk);
     if (results.length > 0) {
       return results[0];
     }
@@ -83,5 +79,4 @@ module.exports = {
   createKatalog,
   editKatalog,
   removeKatalog,
-  searchKatalog,
 };
